# app/src-tauri/resources/utubeholic.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "UtubeHolic Local Server API",
    "version": "1.0.0",
    "description": "Local-only Flask server used by UtubeHolic Chrome extension and the GUI app. Base URL is bound to 127.0.0.1."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5000"
    }
  ],
  "tags": [
    {
      "name": "downloads"
    },
    {
      "name": "watched"
    },
    {
      "name": "ops"
    }
  ],
  "paths": {
    "/download": {
      "post": {
        "tags": [
          "downloads"
        ],
        "summary": "Queue a download task",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadQueuedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/download_action": {
      "post": {
        "tags": [
          "downloads"
        ],
        "summary": "Control a download task",
        "description": "cancel=hard cancel; retry=re-request from scratch; delete=remove from list (optionally with files).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadActionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/status": {
      "get": {
        "tags": [
          "ops"
        ],
        "summary": "Get server status",
        "responses": {
          "200": {
            "description": "Status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          }
        }
      }
    },
    "/logs": {
      "get": {
        "tags": [
          "ops"
        ],
        "summary": "Get recent logs",
        "parameters": [
          {
            "name": "max_lines",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 2000,
              "default": 200
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Logs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/openapi.json": {
      "get": {
        "tags": [
          "ops"
        ],
        "summary": "Get OpenAPI spec",
        "responses": {
          "200": {
            "description": "OpenAPI JSON"
          }
        }
      }
    },
    "/sync_watched": {
      "post": {
        "tags": [
          "watched"
        ],
        "summary": "Upsert watched records into SQLite",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncWatchedRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncWatchedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/watched_export": {
      "get": {
        "tags": [
          "watched"
        ],
        "summary": "Export watched rows (paged)",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 20000,
              "default": 5000
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchedExportResponse"
                }
              }
            }
          }
        }
      }
    },
    "/watched_count": {
      "get": {
        "tags": [
          "watched"
        ],
        "summary": "Get watched row count",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchedCountResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "ok",
          "error"
        ]
      },
      "Cookie": {
        "type": "object",
        "properties": {
          "domain": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "value": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "secure": {
            "type": "boolean"
          },
          "expirationDate": {
            "type": "number"
          }
        },
        "additionalProperties": true
      },
      "DownloadRequest": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "cookies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Cookie"
            }
          }
        },
        "required": [
          "url"
        ],
        "additionalProperties": true
      },
      "DownloadQueuedResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "task_id": {
            "type": "integer"
          }
        },
        "required": [
          "ok",
          "task_id"
        ]
      },
      "DownloadActionRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "action": {
            "type": "string",
            "enum": [
              "cancel",
              "retry",
              "delete"
            ]
          },
          "delete_files": {
            "type": "boolean",
            "default": false
          }
        },
        "required": [
          "id",
          "action"
        ],
        "additionalProperties": true
      },
      "DownloadItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "url": {
            "type": "string"
          },
          "video_id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "created_at": {
            "type": "integer"
          },
          "started_at": {
            "type": "integer"
          },
          "finished_at": {
            "type": "integer"
          },
          "percent": {
            "type": "number"
          },
          "speed": {
            "type": "number"
          },
          "eta": {
            "type": "integer"
          },
          "downloaded_bytes": {
            "type": "number"
          },
          "total_bytes": {
            "type": "number"
          },
          "format": {
            "type": "string"
          },
          "uploader": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          },
          "files": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "error": {
            "type": "string"
          }
        },
        "additionalProperties": true
      },
      "DownloadActionResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "action": {
            "type": "string"
          },
          "task_id": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "deleted": {
            "type": "boolean"
          },
          "delete_files": {
            "type": "boolean"
          },
          "deleted_paths": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "ok"
        ]
      },
      "Progress": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          },
          "percent": {
            "type": "number"
          },
          "speed": {
            "type": "number"
          },
          "eta": {
            "type": "integer"
          },
          "downloaded_bytes": {
            "type": "number"
          },
          "total_bytes": {
            "type": "number"
          },
          "error": {
            "type": "string"
          }
        },
        "additionalProperties": true
      },
      "StatusResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "queue_size": {
            "type": "integer"
          },
          "progress": {
            "$ref": "#/components/schemas/Progress"
          },
          "downloads": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DownloadItem"
            }
          },
          "sqlite_path": {
            "type": "string"
          },
          "download_dir": {
            "type": "string"
          }
        },
        "required": [
          "ok"
        ]
      },
      "LogsResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "lines": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "ok",
          "lines"
        ]
      },
      "WatchedRecord": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "ts": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "ts"
        ]
      },
      "SyncWatchedRequest": {
        "type": "object",
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WatchedRecord"
            }
          }
        },
        "required": [
          "records"
        ],
        "additionalProperties": true
      },
      "SyncWatchedResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "received": {
            "type": "integer"
          },
          "inserted": {
            "type": "integer"
          },
          "updated": {
            "type": "integer"
          },
          "ignored": {
            "type": "integer"
          },
          "count": {
            "type": "integer"
          },
          "sqlite_path": {
            "type": "string"
          }
        },
        "required": [
          "ok",
          "received",
          "inserted",
          "updated",
          "ignored"
        ]
      },
      "WatchedCountResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "count": {
            "type": "integer"
          },
          "sqlite_path": {
            "type": "string"
          }
        },
        "required": [
          "ok",
          "count"
        ]
      },
      "WatchedExportResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "page": {
            "type": "integer"
          },
          "page_size": {
            "type": "integer"
          },
          "returned": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "has_more": {
            "type": "boolean"
          },
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WatchedRecord"
            }
          },
          "sqlite_path": {
            "type": "string"
          }
        },
        "required": [
          "ok",
          "page",
          "page_size",
          "returned",
          "total",
          "has_more",
          "records"
        ]
      }
    }
  }
}